will restrucure addToDB

+++ MSDiskManager/ViewModels/DirectoryViewModel.cs
@@ -1,8 +1,10 @@
 ﻿#nullable enable
 using MSDiskManager.Dialogs;
+using MSDiskManager.Helpers;
 using MSDiskManagerData.Data.Entities;
 using MSDiskManagerData.Data.Entities.Relations;
 using MSDiskManagerData.Data.Repositories;
+using MSDiskManagerData.Helpers;
 using Nito.AsyncEx;
 using System;
 using System.Collections.Generic;
@@ -14,6 +16,8 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Media;
 
 namespace MSDiskManager.ViewModels
 {
@@ -23,12 +27,18 @@ namespace MSDiskManager.ViewModels
         private List<FileViewModel> files = new List<FileViewModel>();
         private List<DirectoryViewModel> children = new List<DirectoryViewModel>();
         private int itemsCount = 0;
-
+        private bool hasRemovedItem = false;
         public virtual List<FileViewModel> Files { get => files; set { files = value; NotifyPropertyChanged("Files"); } }
         public virtual List<DirectoryViewModel> Children { get => children; set { children = value; NotifyPropertyChanged("Children"); } }
         public override IconType IconType => IconType.Directory;
         public int ItemsCount { get => itemsCount; set { itemsCount = value; NotifyPropertyChanged("ItemsCount"); } }
+        public void RemoveChild(BaseEntityViewModel entity)
+        {
 
+            if (entity is FileViewModel) files.Remove((FileViewModel)entity);
+            else children.Remove((DirectoryViewModel)entity);
+            hasRemovedItem = true;
+        }
 
         public bool StartRemoving()
         {
@@ -44,7 +54,16 @@ namespace MSDiskManager.ViewModels
             Files.ForEach(f => f.ShouldRemove = false);
         }
 
-
+        public void FreeResources()
+        {
+            foreach (var f in files) f.FreeResources();
+            foreach (var c in children) c.FreeResources();
+        }
+        public void ResetResources()
+        {
+            Files.ForEach(f => f.ResetResources());
+            children.ForEach(c => c.ResetResources());
+        }
         public async Task<bool> AddToDb(bool move,
             Action<BaseEntityViewModel?, CopyMoveEventType> callback,
             PauseTokenSource pauses, CancellationTokenSource cancels, bool bypass = false, bool dbOnly = false)
@@ -71,15 +90,26 @@ namespace MSDiskManager.ViewModels
                     this.Path = updated.Path;
                 }
                 var result = true;
-                foreach (var file in files)
+                if (Globals.IsNotNullNorEmpty(files))
                 {
-                    result &= await file.AddToDb(move, callback, pauses, cancels);
+                    var success = await new FileRepository().AddFiles(
+                    files.Select(f => f.FileEntity).ToList(), async (f, e) =>
+                     {
+                         pauses.IsPaused = true;
+                         var diag = new CopyMoveErrorDialog($"FILE: [{f.OldPath}]\n{e.Message}", pauses, cancels, async () => result = await AddToDb(move, callback, pauses, cancels, true), (e)=> { });
+                         diag.ShowDialog();
+                         await pauses.Token.WaitWhilePausedAsync();
+                         return cancels.IsCancellationRequested ? 2 : (pauses.IsPaused ? 1 : 0);
+                     }, move: move
+                    );
+                    if (cancels.IsCancellationRequested) return false;
+                    files.ForEach(async f => await Task.Run(() => callback(f, f.Id == null ? CopyMoveEventType.Skip : CopyMoveEventType.Success)));
                 }
                 foreach (var child in children)
                 {
                     result &= await child.AddToDb(move, callback, pauses, cancels);
                 }
-                if (result && move)
+                if (result && move && !hasRemovedItem)
                 {
                     Directory.Delete(this.OriginalPath);
                 }
@@ -93,7 +123,7 @@ namespace MSDiskManager.ViewModels
                 pauses.IsPaused = true;
                 bool result = false;
                 var msg = $"Directory :[{OriginalPath}]\n{ioe.Message}";
-                var diag = new CopyMoveErrorDialog(msg , pauses, cancels, async () => result = await AddToDb(move, callback, pauses, cancels, true), (et) => callback(null, et));
+                var diag = new CopyMoveErrorDialog(msg, pauses, cancels, async () => result = await AddToDb(move, callback, pauses, cancels, true), (et) => callback(null, et));
                 diag.ShowDialog();
                 return result;
             }
@@ -126,6 +156,7 @@ namespace MSDiskManager.ViewModels
             Parent = Parent?.DirectoryEntity,
             IsHidden = IsHidden,
             ParentId = Parent?.Id,
+            OldPath = OriginalPath,
         };
         public int AllItemsCountRecursive
         {
@@ -191,5 +222,22 @@ namespace MSDiskManager.ViewModels
         public override bool IsHidden { get => base.IsHidden; set { files.ForEach(f => f.IsHidden = value); children.ForEach(c => c.IsHidden = value); base.IsHidden = value; } }
 
         public override bool IgnoreAdd { get => base.IgnoreAdd; set { files.ForEach(f => f.IgnoreAdd = value); children.ForEach(c => c.IgnoreAdd = value); base.IgnoreAdd = value; } }
+
+        public override ImageSource Image => "/images/directory.png".AsUri();
+
+        public override double ImageWidth => 50;
+
+        public override object TooltipContent
+        {
+            get
+            {
+                if (content != null) return content;
+                var img = new Image();
+                img.Source = Image;
+                img.MaxWidth = 50;
+                content = img;
+                return img;
+            }
+        }
     }
 }
